files .h & Makefile done, fixe last francinnette bug

+++ ft_strlcat.c
@@ -6,32 +6,31 @@
 /*   By: acombier <acombier@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/11/09 13:07:32 by acombier          #+#    #+#             */
-/*   Updated: 2025/11/11 12:52:12 by acombier         ###   ########.fr       */
+/*   Updated: 2025/11/11 14:06:15 by acombier         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-unsigned int	ft_strlcat(char	*dest, char *src, unsigned int size)
+size_t	ft_strlcat(char *dest, char *src, unsigned int size)
 {
-	int	i;
-	unsigned	int j;
-	unsigned	int len;
+	int				i;
+	unsigned int	j;
+	unsigned int	len;
 
 	i = 0;
 	j = 0;
 	len = 0;
-
-	while(dest[i])
+	while (dest[i])
 	{
 		len++;
 		i++;
 	}
-	if(size < len || size == 0)
+	if (size < len || size == 0)
 	{
 		return (0);
 	}
-	while(j < (size - len - 1))
+	while (j < (size - len - 1))
 	{
 		dest[i] = src[i];
 		j++;
@@ -40,6 +39,3 @@ unsigned int	ft_strlcat(char	*dest, char *src, unsigned int size)
 	dest[i] = '\0';
 	return (len + j);
 }
-
-
-
